Consolidate postgres schema code
