+++ lib/postgres/debezium/data_type.go
@@ -1,53 +1,8 @@
 package debezium
 
 import (
+	"github.com/artie-labs/reader/lib/postgres/schema"
 	"github.com/artie-labs/transfer/lib/debezium"
-)
-
-type DataType int
-
-var (
-	TextBasedColumns = []string{
-		"xml",
-		"cidr",
-		"macaddr",
-		"macaddr8",
-		"inet",
-		"int4range",
-		"int8range",
-		"numrange",
-		"daterange",
-		"tsrange",
-		"tstzrange",
-	}
-)
-
-const (
-	InvalidDataType DataType = iota
-	VariableNumeric
-	Money
-	Numeric
-	Bit
-	Boolean
-	TextThatRequiresEscaping
-	Text
-	Interval
-	Array
-	HStore
-	Float
-	Int16
-	Int32
-	Int64
-	UUID
-	UserDefinedText
-	JSON
-	Timestamp
-	Time
-	Date
-	// PostGIS
-	Point
-	Geometry
-	Geography
 )
 
 type Result struct {
@@ -55,91 +10,91 @@ type Result struct {
 	Type         string
 }
 
-func (d DataType) ToDebeziumType() Result {
+func ToDebeziumType(d schema.DataType) Result {
 	switch d {
-	case Geography:
+	case schema.Geography:
 		return Result{
 			DebeziumType: string(debezium.GeographyType),
 			Type:         "struct",
 		}
-	case Geometry:
+	case schema.Geometry:
 		return Result{
 			DebeziumType: string(debezium.GeometryType),
 			Type:         "struct",
 		}
-	case Point:
+	case schema.Point:
 		return Result{
 			DebeziumType: string(debezium.GeometryPointType),
 			Type:         "struct",
 		}
-	case VariableNumeric:
+	case schema.VariableNumeric:
 		return Result{
 			DebeziumType: string(debezium.KafkaVariableNumericType),
 			Type:         "struct",
 		}
-	case Money, Numeric:
+	case schema.Money, schema.Numeric:
 		return Result{
 			DebeziumType: string(debezium.KafkaDecimalType),
 		}
-	case Boolean, Bit:
+	case schema.Boolean, schema.Bit:
 		return Result{
 			Type: "boolean",
 		}
-	case Text, UserDefinedText, TextThatRequiresEscaping:
+	case schema.Text, schema.UserDefinedText, schema.TextThatRequiresEscaping:
 		return Result{
 			Type: "string",
 		}
-	case Interval:
+	case schema.Interval:
 		return Result{
 			DebeziumType: "io.debezium.time.MicroDuration",
 			Type:         "int64",
 		}
-	case Array:
+	case schema.Array:
 		return Result{
 			Type: "array",
 		}
-	case Float:
+	case schema.Float:
 		return Result{
 			Type: "float",
 		}
-	case Int16:
+	case schema.Int16:
 		return Result{
 			Type: "int16",
 		}
-	case Int32:
+	case schema.Int32:
 		return Result{
 			Type: "int32",
 		}
-	case Int64:
+	case schema.Int64:
 		return Result{
 			Type: "int64",
 		}
-	case UUID:
+	case schema.UUID:
 		return Result{
 			DebeziumType: "io.debezium.data.Uuid",
 			Type:         "string",
 		}
-	case JSON:
+	case schema.JSON:
 		return Result{
 			DebeziumType: "io.debezium.data.Json",
 			Type:         "string",
 		}
-	case Time:
+	case schema.Time:
 		return Result{
 			DebeziumType: string(debezium.Time),
 			Type:         "int32",
 		}
-	case Date:
+	case schema.Date:
 		return Result{
 			DebeziumType: string(debezium.Date),
 			Type:         "int32",
 		}
-	case HStore:
+	case schema.HStore:
 		return Result{
 			DebeziumType: "",
 			Type:         "map",
 		}
-	case Timestamp:
+	case schema.Timestamp:
 		return Result{
 			DebeziumType: string(debezium.Timestamp),
 			// NOTE: We are returning string here because we want the right layout to be used by our Typing library
